+++ classes/suggested-tasks/class-tasks-manager.php
@@ -33,6 +33,7 @@ use Progress_Planner\Suggested_Tasks\Providers\Remove_Terms_Without_Posts;
 use Progress_Planner\Suggested_Tasks\Providers\Update_Term_Description;
 use Progress_Planner\Suggested_Tasks\Providers\Unpublished_Content;
 use Progress_Planner\Suggested_Tasks\Providers\Collaborator;
+use Progress_Planner\Suggested_Tasks\Providers\Select_Timezone;
 
 /**
  * Tasks_Manager class.
@@ -76,6 +77,7 @@ class Tasks_Manager {
 			new Update_Term_Description(),
 			new Unpublished_Content(),
 			new Collaborator(),
+			new Select_Timezone(),
 		];
 
 		// Add the plugin integration.